New file structure for neural networks
