+++ sbi/neural_nets/build_functions/categorial.py
