+++ sbi/neural_nets/embedding_nets/__init__.py
@@ -1,0 +1,5 @@
+from sbi.neural_nets.embedding_nets.cnn import CNNEmbedding
+from sbi.neural_nets.embedding_nets.fully_connected import FCEmbedding
+from sbi.neural_nets.embedding_nets.permutation_invariant import (
+    PermutationInvariantEmbedding,
+)
