+++ sbi/neural_nets/estimators/shape_handling.py
@@ -1,3 +1,6 @@
+# This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
+# under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+
 import torch
 from torch import Tensor
